Various minor improvements - change to json with comments. Flattened json
Switched from pure JSON to JSON with comments to permit some instructions to be added to the JSON file.
Decided that the structure of JSON to match the structure of the GUI was too complicated. Therefore flattened the JSON structure.
Changed string formatting to the f-string system - Python 3.6
Changed the configuration system to a class with an access function.
Changed states of state machine to enum class.