Added check for versions of firmware in pickit3
I believe that this step is necessary for a new pickit3 because we use a different firmware for the default. Also, if the user inadvertently runs MPLAB, it will replace this later firmware with an earlier one which does not support pic16F15 and then report that it does not support pic16F15. I believe that this step replaces the old firmware again.