Changed picky-py to accept the name of the product to be programmed
Picky_py now takes one argument on the launching command line.
That must contain the name of the product about to be programmed.
That name will be used for the name of the upload folder as opposed to the older "fixed" name of "upload". Different products will therefore have different upload folders. Added .desktop file for linux and an example xpm file for a product called "Flicker"